app/requirements.txt
black~=22.3.0
cryptography~=37.0.2
Django~=4.0.1
django-extensions~=3.1.5
django-phonenumber-field[phonenumbers]~=6.3.0
django-timezone-field==5.0
djangorestframework~=3.13.1
drf-jwt~=1.19.2
drf-spectacular==0.22.1
flake8~=4.0.1
isort~=5.10.1
markdown~=3.4.1
psycopg2-binary~=2.9.3
pydot~=1.4.2
pyparsing~=3.0.9
pytest~=6.2.5
pytest-cov~=3.0.0
pytest-django~=4.5.2
pytz==2022.1
tzdata==2022.1
